requirements.txt
opencv-python==4.8.0.74
tensorflow-gpu==2.12.0
numpy==1.24.3
pandas==2.0.3
moviepy==1.0.3
ffmpeg-python==0.2.0
pynvml==11.505.58
psutil==5.9.8
